+++ src/components/DatePicker/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-11-28: Support `date` and `onDateChange` props to make the component
   controllable.
 - 2025-12-12: Use `IbsBase` rather than deprecated `InputBase` component.
+- 2026-02-13: Use `IbsBaseRootProps` instead of `ComponentProps<typeof
+  IbsBase>` for a clearer root/style contract.