+++ src/components/FileInput/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-11-21: Initial version.
 - 2025-12-08: Use `IbsBase` rather than the deprecated `InputBase` component.
+- 2026-02-13: Use `IbsBaseStyleProps` in `FileInputProps` and keep custom
+  `onChange` typing to avoid root event type pollution.