+++ src/components/Input/CHANGELOG.md
@@ -17,3 +17,5 @@
 - 2025-12-07: Use `IbsBase` component instead of `InputBase` component.
 - 2025-12-18: Use `useOptionalFieldContext` hook to get the field context for
   the `id` prop's default value. Support `status` prop to set the error state.
+- 2026-02-13: Use `IbsBaseStyleProps` in `InputProps` to avoid DOM event type
+  pollution from `IbsBase` root props.