+++ src/components/MultiSelect/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-11-06: Initial version.
 - 2025-12-16: Use `IbsBase` as the base component, etc.
+- 2026-02-13: Use `IbsBaseRootProps` in `MultiSelect` and trigger context
+  types instead of `ComponentProps<typeof IbsBase>`.